upload py notebook with more analysis
