--- machine-learning/old/practice/tf-vox-gender.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-
-def extract_features(raw_sound):
-    print('Extracting...')
