+++ machine-learning/old/tf-practice/tf-clothes.py
@@ -1,0 +1,72 @@
+import tensorflow as tf
+from tensorflow import keras
+import numpy as np
+from tensorflow import keras
+from sklearn.model_selection import train_test_split
+import matplotlib.pyplot as plt
+
+
+def preprocess(x, y):
+    x = tf.cast(x, tf.float32) / 255.0
+    y = tf.cast(y, tf.float64)
+
+    return x, y
+
+
+def create_dataset(xs, ys, n_classes=10):
+    ys = tf.one_hot(ys, depth=n_classes)
+    return tf.data.Dataset.from_tensor_slices((xs, ys)) \
+        .map(preprocess) \
+        .shuffle(len(ys)) \
+        .batch(128)
+
+
+(x_train, y_train), (x_val, y_val) = keras.datasets.fashion_mnist.load_data()
+
+# (x_all, y_all), (x_val, y_val) = keras.datasets.fashion_mnist.load_data()
+# x_train, x_test, y_train, y_test = train_test_split(
+#     x_all, y_all, test_size=0.1)
+
+train_dataset = create_dataset(x_train, y_train)
+val_dataset = create_dataset(x_val, y_val)
+
+model = keras.Sequential([
+    keras.layers.Reshape(target_shape=(28 * 28,), input_shape=(28, 28)),
+    keras.layers.Dense(units=256, activation='relu'),
+    keras.layers.Dense(units=192, activation='relu'),
+    keras.layers.Dense(units=128, activation='relu'),
+    keras.layers.Dense(units=10, activation='softmax')
+])
+
+
+model.compile(optimizer='adam',
+              loss=tf.losses.CategoricalCrossentropy(from_logits=True),
+              metrics=['accuracy'])
+
+history = model.fit(
+    train_dataset.repeat(),
+    epochs=10,
+    steps_per_epoch=500,
+    validation_data=val_dataset.repeat(),
+    validation_steps=2
+)
+
+
+predictions = model.predict(x_val)
+
+predicted = list()
+expected = list()
+for i in range(len(predictions)):
+    predicted.append(np.argmax(predictions[i]))
+
+m = keras.metrics.Accuracy()
+_ = m.update_state(predicted, y_val)
+print(m.result().numpy())
+
+# plt.plot(history.history['loss'])
+# plt.plot(history.history['val_loss'])
+# plt.title('Model loss')
+# plt.ylabel('Loss')
+# plt.xlabel('Epoch')
+# plt.legend(['Train', 'Test'], loc='upper left')
+# plt.show()
