+++ machine-learning/old/tf-practice/tf-vox-gender.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+
+def extract_features(raw_sound):
+    print('Extracting...')
